\name{epclaims}
\alias{epclaims}
\docType{data}
\title{Episodic Claims}
\description{
Fictitious claims data used to illustrate \code{kdpec}.
}
\format{
  \describe{
    \item{\code{PatientID}}{Number to identify each patient.}
    \item{\code{ClaimNumber}}{A number to identify each claim.}
    \item{\code{Diagnosis}}{A character (\code{A} or \code{B}) identifying two fictitious diagnoses.}
    \item{\code{ServiceStart}}{The claim's service start date.}
    \item{\code{ServiceEnd}}{The claim's service end date. }
  }
}
\keyword{datasets}
