\name{ipadmits}
\alias{ipadmits}
\docType{data}
\title{Inpatient Admissions Dataset}
\description{A dataset of fictional inpatient admissions designed to demonstrate several functions in the \code{phalen} package. }
\format{
  \describe{
    \item{\code{HospID}}{Hospital unique identifier.}
    \item{\code{Age}}{Age of patient.}
    \item{\code{isReadmission}}{Readmissions are coded as \code{1}.}
    \item{\code{cost}}{Per day cost of inpatient stay.}
  }
}
\keyword{datasets}
